refactor(conformance): rename _contract_fields.py to avoid collision
suite/checks/prescriptions/_contract_fields.py already exists (unrelated
P011/P012/P015 check functions). Different package, so no actual import
collision, but the identical basename is confusing in a fuzzy file search.
Rename the new neutral module to _entrypoint_contract_fields.py.
