+++ packages/conformance/conformance/suite/checks/_entrypoint_contract_fields.py
