+++ packages/conformance/tests/test_sdk_contract_mixins.py
@@ -4,7 +4,7 @@
 ``application_sdk.contracts.base.Input`` / ``Output`` / ``PublishInputMixin``,
 because those classes live outside the scanned repo when a checker (or the
 ledger generator) runs against a consumer app — see ``resolve_contract_fields``
-in the neutral ``_contract_fields`` module.
+in the neutral ``_entrypoint_contract_fields`` module.
 
 This test keeps the registry honest as the SDK evolves: it locates the real
 ``application_sdk`` package (a ``test`` extra dependency of this package),
@@ -21,7 +21,7 @@ import importlib.util
 from pathlib import Path
 
 import pytest
-from conformance.suite.checks._contract_fields import _iter_fields
+from conformance.suite.checks._entrypoint_contract_fields import _iter_fields
 from conformance.suite.checks._sdk_contract_mixins import (
     SDK_CONTRACT_BASE_FIELDS,
     SdkField,
